+++ libs/base/src/data/bit/field/mod.rs
@@ -1,4 +1,4 @@
-// devela::data::bit::field
+// devela_base::data::bit::field
 //
 //!
 //