+++ libs/base/src/data/bit/wrapper/mod.rs
@@ -1,4 +1,4 @@
-// devela::data::bit::wrapper
+// devela_base::data::bit::wrapper
 //
 //!
 //
@@ -15,7 +15,8 @@ mod tests;
 /// [`i8`], [`i16`], [`i32`], [`i64`], [`i128`], [`isize`],
 /// [`u8`], [`u16`], [`u32`], [`u64`], [`u128`] and [`usize`].
 ///
-/// See also [`BitOps`][super::BitOps] for the corresponding trait.
+/// See also [`BitOps`] for the related trait.
+#[doc = "\n\n[`BitOps`]: crate::BitOps"] // FIXME
 ///
 /// [`i8`]: Self#implementation-for-i8
 /// [`i16`]: Self#implementation-for-i16